+++ jvm/src/test/scala/TranscodingSpec.scala
@@ -9,7 +9,7 @@ class TranscodingSpec extends FlatSpec {
 
 
 
-  "A code point transcoder object"  should "derive a Unicode Greek string when instantiated with an ASCII string" in  {
+  "A code point transcoder object"  should "derive a Unicode Greek string when instantiated with an ASCII string" in  pending /*{
     val lgs = LiteraryGreekString("mh=nin")
     val expected = Normalizer.normalize("μῆνιν", Normalizer.Form.NFC)
     assert(lgs.ascii == "mh=nin")
@@ -32,15 +32,27 @@ class TranscodingSpec extends FlatSpec {
 
   it should "swap combining smooth breathing when it precedes a word" in {
     //ὡς ἐμοίχευεν ̓Ερατοσθένης τὴν γυναῖκα
-      val eraucode = "Ἐρατοσθένης"
+    val eraucode = "Ἐρατοσθένης"
 
-      // Generate from literaryUcodeOf function:
-      val lgs = LiteraryGreekString(literaryUcodeOf(eraucode))
-      assert(lgs.ascii == "*e)ratosqe/nhs")
-      // Generate directly from String value
-      assert(lgs.ascii == LiteraryGreekString(eraucode).ascii)
-    
+    // Generate from literaryUcodeOf function:
+    val lgs = LiteraryGreekString(literaryUcodeOf(eraucode))
+    assert(lgs.ascii == "*e)ratosqe/nhs")
+    // Generate directly from String value
+    assert(lgs.ascii == LiteraryGreekString(eraucode).ascii)
+
+  }*/
+
+
+
+/*
+  it should "enable nice sorting of a vector" in {
+    val wds = Vector ("ὡς", "ἐμοίχευεν", " ̓Ερατοσθένης", "τὴν", "γυναῖκα")
+    val lgs = for (wd <- wds) yield (LiteraryGreekString(wd))
+    val expectedWithCase =  Vector("Ἐρατοσθένης", "γυναῖκα", "ἐμοίχευεν", "τὴν", "ὡς")
+    assert(expectedWithCase == lgs.sortWith(_ < _).map(_.ucode))
+    println(lgs.map(_.toLower).sortWith(_ < _).map(_.ucode))
+
   }
-
+*/
 
 }